Initial working version
Further testing is required, but the current version is able to compile the sample application without errors.

The files have been modified to make OSAL compatible with the transpiler. In addition, the functions to install the interrupt handlers for RTEMS have been added.

The unit tests have to be modified to adapt them to the new implementation.