Added support for throw and reboot statements
Actions now do not return a Result type, but return an enumerated type that is outside the Termina language. This new enumerated type can have different values depending on what was the last statement executed by the function.

We have removed the Result type and defined the new type __action_result_t. The latter is the new type of the return values of the actions.