feat: added emulated keyboard interrupt in POSIX system support
In this new version of the software, we provide support for an emulated interrupt in POSIX systems. The OSAL is capable of generating an interrupt (active by level) every time there is new data on the keyboard. In addition, we have added a new system call, read(), which allows data to be read from standard input in a non-blocking manner. This new implementation will allow the behavior of an interrupt-based system to be emulated and more realistic POSIX demonstrators to be implemented.

In addition, a minor bug has been fixed in the RTEMS implementation, where a variable did not have the correct type. The system did not fail because, in the LEON system, the size of the previous type (size_t) was the same as the correct size (uint32_t).

Finally, we have fixed the binaries for LEON to be compiled with the soft-float flag. In the future, it will be necessary to change this to make it parameterizable.